requirements.txt
python-redmine==2.0.2
Django==1.11
django-braces==1.11.0
